src/requirements.txt
python_dotenv==0.21.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
python_dotenv==0.21.0

fastapi==0.92.0
fastapi-jwt-auth==0.2.0
fastapi-users==10.4.0
uvicorn==0.20.0

PyJWT==2.6.0
python-decouple==3.7